Move import/export data controls onto Settings with erase.
Keep backup export/import next to erase-all under Wallet data, and remove the duplicate backup panel from Accounts.
